PR-02: extract dbus_naming.rs
Behaviour-preserving relocation:
- errors.rs: gained DbusSuffixError + Display/Error impls (was in
  main.rs near the top with the bus-name helpers).
- dbus_naming.rs (new): sanitize_dbus_suffix, derive_default_dbus_suffix,
  resolve_dbus_suffix, effective_dbus_name, is_daemon_bus_name.

main.rs: 7743 -> 7642 LOC. Added mod/use lines and extended the
#[cfg(test)] re-export glob.

Verification: cargo build clean; cargo test --bin kanata-switcher
261 passed / 0 failed.
